+++ spike/codehike/src/dependencies.ts
@@ -1,3 +1,0 @@
-/// <reference types="@esri/calcite-components/types/react" />
-/// <reference types="mdx" />
-export {};